autosharder: set limits on the calculated overhead
android_browsertests on android-x64-rel was not sharded enough due to a
bad overhead calculation. It looks like the suite was just stuck in an
unforunate set of shardings which made the overhead look far worse than
it really was (~10mins). To address cases like this, clamp the overhead
calculation to approximately the current min and max overheads. In the
long term we should use more than 1 shard configuration.

Bug: None
Change-Id: I5470efc7afcd6aeaa8a9cc28c09ec207cb3bb57c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6374266
Reviewed-by: Benjamin Joyce (Ben) <bjoyce@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1435477}
NOKEYCHECK=True
GitOrigin-RevId: def11e9193c1132a583f59f9180060b512f47087