ozone/wayland: Tab drag interactive ui test fixes for mutter
Update mutter to include fixes made on the compositor side:
https://crrev.com/c/6318620 https://crrev.com/c/6371819

Add a couple fixes to tab_drag_controller_interactive_uitest.cc:
- Prevent `BrowserChangeWaiter` from not running the closure in case
quit is called another time before the closure gets to run.
- In `DragAllToSeparateWindow` add a similar step as
`DragToSeparateWindow` to ensure the tabs are dragged to the end of the
tab strip.

Use 1920x1200 display size in xvfb.py when running tests with mutter.
This ensures the tab drag tests request a valid size which is more than
the minimum size of a chrome window.

Update the test filter based on tests that are now passing with the
mutter changes, and add 2 new exclusions for tests that are flaky or
broken on mutter.

See the attached bug for more details.

Bug: 391907155

Change-Id: Ieb46a7209adfcd2a11ff04e8cb738b71b6c3e34a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6310437
Reviewed-by: David Pennington <dpenning@chromium.org>
Reviewed-by: Max Ihlenfeldt <max@igalia.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Orko Garai <orko@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1436130}
NOKEYCHECK=True
GitOrigin-RevId: 30975045b7f83d3c9450f1edf86f457c227b4539