[NVS-httpcache] update filed trial config
We're going to run the study by combining the two flags.

This CL also has to disable the two flags for
HTTPCacheSearchPreloadUnifiedBrowserTest.BackwardHitHttpCache
because we need to update some conditions to make it pass.
There is an ongoing CL to update the conditions, but it gets stuck
due to the test infra flakiness.
https://crrev.com/c/6383241

Bug: 388557183
Change-Id: I6d6a4f8d8dfb4855472a74059016a22b58491346
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6397504
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1440097}
NOKEYCHECK=True
GitOrigin-RevId: 7dbadc4dc1f37252005b01f466453696d222f6cf