Add fieldtrial testing config for "AndroidBookmarkBar".
This CL adds a fieldtrial testing config for "AndroidBookmarkBar" in
anticipation of an experiment that will turn on the Android bookmark bar
feature for the desktop form factor on Android.

Note that "Appearance" settings are also turned on as that is the new
preference screen which hosts the bookmark bar visibility toggle.

Bug: 347632437
Change-Id: I05cfff6f99767070f17e6aa7e2e3032da74741b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6422712
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Feels: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441234}
NOKEYCHECK=True
GitOrigin-RevId: 60dff5e4bc68dd051d2aa7660d56fa87ec037342