autosharder: parameterize tables
The first step to reworking autosharder is to introduce tests to verify
the SQL is doing what we expect when it is refactored/reworked. This
will require the scripts to be able to be run against mock tables in a
controlled dataset which we can do by parameterizing the table names.

Bug: 407846444
Change-Id: I64db6077f12dfbc48aa024cde936a43096178d5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6423231
Feels: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1441669}
NOKEYCHECK=True
GitOrigin-RevId: 79581a0e756fc687c43a6f237baeef3661ae76bb