[Extensions] Collapse app menu submenu when no manageable extensions
This changes the "extensions" submenu in the main menu to be a single
"find extensions" entry when no extensions would be shown on the
chrome://extensions page.

In this case, the "extensions" submenu would otherwise only contain
two items:
 - Manage extensions (leads to a blank page)
 - Visit Chrome Web Store (confusing to new users)

A separate browser command and UMA metrics are added for the new menu
item in order to be able to see the difference in click-through rates
between the two options.

This change is behind a flag that is disabled by default.

NO_IFTTT=Commands are not gated on fenced frame, etc.

Bug: 407733524
Change-Id: Ifb5556d20d2f8735fb13fe78c0afb1fb4dc4499e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6418443
Auto-Submit: Dana Fried <dfried@chromium.org>
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Feels: Dana Fried <dfried@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441714}
NOKEYCHECK=True
GitOrigin-RevId: edf18f987a2137a01acc52f4602a584c2620c74c