Add pixel test coverage for cc::Layer::BackdropFilterQuality.
Apparently when we added this function a while ago we didn't add any
tests for it. This CL adds some basic pixel tests. It looks like the
software compositor ignores this parameter altogether (i.e., it's
unimplemented), and the difference on GPU is very subtle, but some
coverage is still better than nothing.

Change-Id: I8060a99e6c73dc44ba24e1ef28955606abb5b79b
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6410276
Reviewed-by: Victor Miura <vmiura@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#1442281}
NOKEYCHECK=True
GitOrigin-RevId: 5be942618ab4c673261132a29a6353d23cc4dafe