Add `ListAccountsUsesBinaryFormat` to `fieldtrial_testing_config.json`
With the enabling of the flag, some tests started failing because the
ListAccounts response was still being formatted using jspb, while the
parsing was done in binary. This fixes the tests by creating the
ListAccounts response in a binary format if the parameter "laf=b64bin"
can be found in the URL. This parameter is set when the flag is enabled.
The old format is still supported.

Bug: 356643207
Change-Id: Iebcb30861f3ac2f835e86b73da9cd18f1e602024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6415691
Commit-Queue: Amelie Schneider <amelies@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Anthi Orfanou <anthie@google.com>
Cr-Commit-Position: refs/heads/main@{#1442627}
NOKEYCHECK=True
GitOrigin-RevId: 6f48889585bc0c191ad6cde42469f886cee46ef1