Add feature to set the spare renderer as the lowest priority on Android
The CL adds a new feature parameter to keep the spare renderer as the
lowest priority. This allows the spare renderer to be always the first
process to be killed during OOM and protects other processes.

The invetigation doc on the OOM issues with the spare renderer can be
found at:
https://docs.google.com/document/d/1tNIQs3YgqK9QczvB5Sivkqa7LasRZV3kzkf9ooBPvy4/edit?usp=sharing&resourcekey=0-mCY9bLZsaeA1bG_FeiRV6w

Bug: 409693027
Change-Id: I887cb01c23a75f19d52d7f289970b069644f0b4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6445591
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Jiacheng Guo <gjc@google.com>
Cr-Commit-Position: refs/heads/main@{#1447026}
NOKEYCHECK=True
GitOrigin-RevId: ea5f0aa1812e7e0d575748b59c85b07ff1e4399d