Add an experiment for unlocking database on close
The lock on windows is using the LockFileEx(...) API. Unfortunately,
the lock is not released on process exit but is 'eventually'
released. Which may leads to race condition on a Chrome restart
(most likely, chrome updates).

This CL is adding an experiment to observe these side-effects.

Documentation:
  https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-lockfileex

"""
If a process terminates with a portion of a file locked or closes a file that has outstanding locks, the locks are unlocked by the operating system. However, the time it takes for the operating system to unlock these locks depends upon available system resources. Therefore, it is recommended that your process explicitly unlock all files it has locked when it terminates. If this is not done, access to these files may be denied if the operating system has not yet unlocked them.
"""

Bug: 369635654
Change-Id: I12215ad7716e95f307b5ed08968302eb9b8a509e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6462574
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447987}
NOKEYCHECK=True
GitOrigin-RevId: 799be5e132eb36551b4bd2b9d1b89d1b1b7e7595