Create three finch groups for feature VSynAlignedPresent
The new finch groups for VSyncAlignedPresent are "AllFrames",
"Animation", and "Interaction".

With VSynAlignedPresent enabled, the frame is presented at the next
VSync only when this frame handles interaction or animation as described
in the finch group name kTargetForVSync. Default is set to "AllFrames"
which applies the feature to all frames.

Remove the old flag kPresentationDelayForInteractiveFrames. It's no
longer needed. This group still has big regressions on guardrail
metrics, although it improves a lot on smoothness metrics.

Bug: 330771325
Change-Id: I292360e0fe4a5c3558fcda96a0720a8c5d554e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6482387
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451162}
NOKEYCHECK=True
GitOrigin-RevId: 66a8068b5e3296e1edffcb183b1e20a1eebd4fc1