Add a finch experiment for fixed memory map size
This CL is adding a finch study that force a fixed size for the amount
of memory used by sqlite for memory map database.

The function ComputeMmapSizeForOpen() returns the appropriate size
of memory map that is considered not corrupt. This is a misconception
of the way sqlite use the mmap size. In practice, it won't map
more than the specified memory but from any offset in the file.

The goal of the experiment is to show that there won't be regression;
neither stability issues.

Bug: 413695811
Change-Id: I48a8ae12083d501850e48324afac1813f5258b3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6488592
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1452019}
NOKEYCHECK=True
GitOrigin-RevId: e7569b35c44cdf63b14441f13cb9c6d5a6e7d6ed