[Crossbench] Show the benchmark runner output if the test fails.
If the test fails, the error message is available in the `benchmark_log.txt` log file. Printing this log to the console displays the error in the Swarming task UI, eliminating the need for users to search the CAS Outputs folder. To prevent overly long UI reports, especially when the '--debug' switch is used and the log file is extensive, only the last 100 lines will be printed.

Fixed: 408967807
Change-Id: I041c25448907f16d2eb8ac8450a1a1246e93193e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6506996
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Cr-Commit-Position: refs/heads/main@{#1455191}
NOKEYCHECK=True
GitOrigin-RevId: 6aa1d1546c60eb00b97293c1533cfcb4c57db2b2