Delay screenshot SendResult and compression
SendResult and compression are expensive operations that can cause
contention with loading operations during navigation.

We don't need a screenshot immediately, but after the user reacts to a
page load, they might decide to do a history navigation gesture.
In other words, we can't wait that long. This CL uses a delay of
400ms, configurable via trial fields.

Screenshot compression can be delayed until the browser reaches
quiescence.

This CL replaces the previous approach of using lower priorities, which
proved not to be enough.

Bug: 393558258
Change-Id: Ibf4092ceb9d7adc5ff265fe1beec3819509924cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6502812
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1455899}
NOKEYCHECK=True
GitOrigin-RevId: aee41bb06af91633fd4668659eb661f68ab85cdb