ozone/wayland: Actually observe windows in XdgForeignWrapper
XdgForeignWrapper implements WaylandWindowObserver but it didn't
register itself with the window manager. So it wasn't getting window
destroyed notification and hence holding onto dangling references to
wayland surfaces, causing a crash for debug builds at the end of the
test IFrameTest.FileChooserInDestroyedSubframe in mutter.

So ensure it is registered to be notified of window removal so that
references to deleted exported surfaces can be cleared.

Also, remove the test from the filter.

Bug: 395595608
Change-Id: I0d80267c0c216a4595711505f08600e45fa3f41e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6512163
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Auto-Submit: Orko Garai <orko@igalia.com>
Commit-Queue: Orko Garai <orko@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1456682}
NOKEYCHECK=True
GitOrigin-RevId: 0c5196946f52a21a6b51628050ce2e39c05a91ce