[ChromeOS] Remove UseDMSAAForTiles from test config
Per http://b/349761862 this feature is known to be broken broken for the
ChromeOS device "steelix". There are also no active studies for this
feature on ChromeOS, so it shouldn't be turned on as part of the
ChromeOS fieldtrial testing config.

Bug: 349761862
Change-Id: I954e33ddbaab7684588629db784da27ebc51c6c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6343427
Auto-Submit: Ross Zwisler <zwisler@google.com>
Commit-Queue: Ross Zwisler <zwisler@google.com>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1457737}
NOKEYCHECK=True
GitOrigin-RevId: 45ee32f8cf0eda5541209dcb23ddc3cd76c0baaa