[owners] Remove skyostil@chromium.org from testing/android/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/skyostil@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ie81de115db37e3bda41146e6425a131385f803ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554003
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461244}
NOKEYCHECK=True
GitOrigin-RevId: 3ae08b8447d390076ee3e3f7d27c431fc09ff985