Deflake Autofill interactive UI tests on the force-accessibility-rel bot
Around 85 Autofill interactive UI tests are flaky/fail on the
linux-blink-web-tests-force-accessibility-rel bot. The reason why is
that ShowAutofillSuggestions assumes that the suggestions will not have
been popped up as a result of giving the input focus. However,
AutofillAgent::HandleFocusChangeComplete shows the suggestions if a
screen reader is enabled or --force-renderer-accessibility is used.

In real-world usage, there does not appear to be any problems happening
even with a screen reader running. Therefore, the easiest way to deflake
these 85 tests is to synthesize an Escape before synthesizing the Down
Arrow if the AXMode is kScreenReader.

Bug: 417978157
Change-Id: Ib02606f32cc6e1bf2f1853aa583b4776e3cb5876
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6549941
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461390}
NOKEYCHECK=True
GitOrigin-RevId: 3b87bfe749e926988a59b9bbbc446956748d7df0