Fix a11y dangling-pointer issue for FindRequestManagerTest.FindInPage_Issue627799
When DestroyAtkObjects() is called, remove the object being destroyed
from the find-in-page results map so that we don't attempt to reference
it later.

Also update accessibility-linux.content_browsertests.filter to remove
one test that no longer exists and another which locally is not flaky.

Bug: 418737570
Change-Id: Ib56a5fdf8a56f14976e22ccccab3a86bac13ca1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6563548
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1462183}
NOKEYCHECK=True
GitOrigin-RevId: 15c3f2c49b64aeb27126a07faf527782c9d1de87