Enable ServiceWorkerStaticRouterRaceNetworkRequestPerformanceImprovement
We have run the experiment for a while. We saw expected improvements
with this feature on the partner website, and don't see any issues so
far. So this CL enables this by default.

With this feature, the browser processes the response data sequentially
for the network request data pipe then for the fetch handler data pipe.
This resulted in some test code changes and error handling tweaks.

Bug: 374606637
Change-Id: I0ee572f68c6dfa231a4ca2229cfc57730c63c209
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6332391
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462520}
NOKEYCHECK=True
GitOrigin-RevId: bec0542839a56e183b0559fffcdc564ca744a730