Fix a11y failure of AutofillInteractiveFencedFrameTest.DeletingFrameClosesPopup/1
This test fails on linux-blink-web-tests-force-accessibility-rel due to
triggering a DCHECK (unknown AXTreeID for the focused tree). This can
occur when nothing has focus. It is possible to obtain this id via the
AXTreeData associated with the root AXPlatformNodeDelegate. Doing so
fixes the failure.

Also triage the remaining interactive UI tests in the filter file for
this bot.

Bug: 418640888
Change-Id: I63d68bd2c5f83420ea486f9ddad8592533228c45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6563466
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1463243}
NOKEYCHECK=True
GitOrigin-RevId: f2c22e3c5d8c136dc14432d1c68123ecd84e66a0