Enable desktop-like renderer process limits for Android.
By setting a soft limit, this enforces RPH to share renderers in the
same way as we do in desktop. As a result of sharing, it becomes
much less likely to hit the upper bound of the system.

Bug: 414894073
Change-Id: Ie1d41db18127e396b5decdbf4590136f740c1212
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6551276
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Auto-Submit: Sangwhan Moon <sxm@google.com>
Cr-Commit-Position: refs/heads/main@{#1464337}
NOKEYCHECK=True
GitOrigin-RevId: 856dd0925f0742665a1a3a924c360bee710e8b62