Implement various solutions for Messages AccessibilityEvent crashes
This CL implements the bulk of the experiment to test sending
AccessibilityEvents in the Messages flow. We try a variety of
approaches since the origin of the crash is unclear, as is the best
possible solution. The experiment has a FeatureParam which is used to
split users into groups that try each of the possibly solutions
separately.

AX-Relnotes: N/A
Bug: 417721160
Change-Id: If733c30eefec83e6dbeb18a015e418c8de34cbd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6568873
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Auto-Submit: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Cr-Commit-Position: refs/heads/main@{#1464441}
NOKEYCHECK=True
GitOrigin-RevId: 8400452ce0c4e19f22283e7f2a629c9a29e594eb