Make chrome ios smoke tests aware of Seed File.
Before this CL, the test only checks Local State, now it checks the seed
file, to make sure that the seed is stored in a file.

To achieve that, it changes the way that the seed is injected into the
test, so when it inject a seed, it writes to both Local State and the
VariationsSeedV1 file.

Also it waits if a write is in progress, to make sure the check happens
after the seed is written.

Bug: 413580167

Change-Id: Iac24ccfda2859afaf86643905d18c1042eea65ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6546896
Reviewed-by: Ernest Nguyen Hung <ernn@google.com>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Commit-Queue: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1464597}
NOKEYCHECK=True
GitOrigin-RevId: 37bab131cdc53f8f56e52c0f0b4d559cc784a666