Introduce kBackgroundUpdateForRegisteredStorageKeys feature
The initial navigation on browser startup tends to be blocked by
checking ServiceWorker registration. Checking the registration of
ServiceWorker runs in the thread pool of the browser process, so the DB
operation itself can be run without being blocked by other busy tasks.
But, to receive the result, a task must be run on the UI thread
currently. Since the UI thread is usually very clouded, receiving
results can be delayed a lot.

This CL introduces kBackgroundUpdateForRegisteredStorageKeys feature to
mitigate this issue. With this feature, it makes it possible to receive
the `registered_storage_keys_` information from the thread pool without
waiting for running the receiving task. To receive the information in
the thread pool from UI thread, this CL introduces StorageSharedBuffer
class.

Bug: 421530699
Change-Id: I8bc703bad94fa0373a80243379f8d828d7ec7a2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6606856
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467981}
NOKEYCHECK=True
GitOrigin-RevId: cdd9f95bd77a0683caf21fb962a8c085b28b0e7a