[A11yPerformance] Reset Accessibility Experiment in Task
The Accessibility experiment was reset potentially during a11y access.
This could cause issues, as it was being done mid-access of the tree.
This change will make the reset be a task, so that it runs when it is
safe to do so.

In addition to that, the flags that conflict with the perf experiment
were also refined, allowing flags that does not interact with web
content to be allowed. For example, filter flags or annotation ones are
allowed, so that the experiment is not stopped too early.

Change-Id: I1cddfd95fa09f5aaa19be337ea5874618e720367
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6575866
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1468339}
NOKEYCHECK=True
GitOrigin-RevId: 14927a86c94de58fe76850ea560426663b1e10f1