Make android-desktop browsertests filters more generic
It turns out that the expectations in
android.desktop.emulator_15.android_browsertests.filter apply to
hw tests as well. So this takes that filter file and duplicates it
in a more generic file name while trimming it down to the expectations
that apply to both emu and hw bots.

In a follow-up CL, we can then take the current
android.desktop.emulator_15.android_browsertests.filter file and
trim it down to emu-only expectations, then point emu bots to use both
that filter and this new one. That way expectations aren't duplicated.
Have to do that in a follow-up since these files are referenced in
the downstream repo too.

Bug: 422768734
Change-Id: Iab16c6251068ac87a8a1fa8d091ac5862ccdea3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6635693
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472173}
NOKEYCHECK=True
GitOrigin-RevId: ddcea4278a424f292d4bb0cde989cb7b69772351