Add mojom to send frames to Boca Ui.
This change is for the Class Tools feature, Spotlight, which allows
teachers to view their students screen. This is facilitated via a
read-only CRD session between the teacher and student devices.
Specifically this change introduces functionality to the CRD client to
consume frames from the student and send those frames to the Boca UI.

Bug: 409613172

Change-Id: Ie559ca94356ac6d781f366279a468bd3002580fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6580544
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: April Zhou <aprilzhou@google.com>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: April Zhou <aprilzhou@google.com>
Cr-Commit-Position: refs/heads/main@{#1473240}
NOKEYCHECK=True
GitOrigin-RevId: 7ce08198debc2a1834dd4490e05f6f6447bfe06a