[a11y] Allow AT detection on page navigation when AXMode is complete
Some platforms may not perfectly signal when an
assistive technology (AT) is active, and may
re-check for ATs after the AXMode changes. If the
AXMode is already configured with kAXModeComplete
(meaning all accessibility features are enabled),
a new AT trying to access APIs won't cause a mode
change because those flags are already present.
This change allows for an AT to be detected and
computed when a page loads, specifically in the
rare scenario where the current mode is
yet. This ensures the algorithm is complete.

ui: :kAXModeComplete but no AT has been identified
Change-Id: Iec4a77fd6b2ab35e4c1f8477f071e03a761487a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641432
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1473778}
NOKEYCHECK=True
GitOrigin-RevId: 2c035ba795eeb1581f4d128b9769fff21bdc6c6b