Add OzonePlatformAutoExternal fieldtrial testing config
This is separate from the "OzonePlatformAuto" config which is currently
used for internal dogfooding only.

Rationale for NOPRESUBMIT: The following false-positive error is raised:
"The following feature(s) were specified in multiple studies:
OverrideDefaultOzonePlatformHintToAuto (in studies
OzonePlatformAutoExternal, OzonePlatformAuto)". The check is unaware
that the configs will be separated based on dogfood/non-dogfood
populations, which is only specified in the Finch config, not in the
public Chromium repo. testing/variations/PRESUBMIT.py allows bypassing
PRESUBMIT for cases like this, based on this docstring:

"Note that a feature may be specified in different studies that do not
overlap. For example, if they specify different platforms. In such a
case, this will not give a warning/error. However, it is possible that
this incorrectly gives an error, as it is possible for studies to have
complex filters (e.g., if they make use of additional filters such as
form_factors, is_low_end_device, etc.). In those cases, the PRESUBMIT
check can be bypassed. Since this will only check for studies that were
changed in this particular commit, bypassing the PRESUBMIT check will
not block future commits."

NOPRESUBMIT=true

Bug: 40083534
Change-Id: I6441f7334bd2967611f8b34d2d5373400c25c1a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6649108
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1474710}
NOKEYCHECK=True
GitOrigin-RevId: 03d404f1a0811470f150577ca7c40071ceb797ff