testing: Reenable EncryptedMedia tests on ChromeOS
As per the discussion in the bug, the EncryptedMedia tests were
re-enabled for linux, and were meant to be re-enabled for chromeos, and
was put on the backburner. This CL is meant to complete that work and
re-enable the tests, and once the failure is not observed, we can mark
the bug as fixed.

Bug: 41466539
Change-Id: I5d76243ce50c0cec223e2c6d1edc1e0785550bcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6652729
Reviewed-by: Feras Aldahlawi <frs@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475037}
NOKEYCHECK=True
GitOrigin-RevId: 3d169d82f5a69759aaf3765f46769973437a2da0