Allow GN analyzer to filter down the `tools/perf/cbb_ref_info` path.
The `tools/perf/.*` is excluded from the GN analyzer which causes to run
all CQ tests. The `tools/perf/cbb_ref_info` has info files which only
are used for the CBB performance testing. We need to allow GN analyzer
to filter the tests for this folder by adding it to the `ignores` list.

Bug: 388894957
Change-Id: I9886a7bf3a952128faad9a0dca9ebc219ceb706b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6653186
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1475067}
NOKEYCHECK=True
GitOrigin-RevId: f22b72f5135cecd05c878b15c4baa6eb1263c2fa