[Autofill] Do not extract checkboxes, radio buttons
This change makes checkboxes and radio buttons as no
not-autofillable elements. As a result, they're no longer
extracted and sent to the browser.

Since checkboxes and radio buttons are filtered in lots of places
in the code anyway -- in particular, they're not autofilled and
they're skipped for the form signature -- the change is expected
to have no impact.

The main changes are
- isAutofillableInputElement in fill_element_inference_util.ts
- ToAutofillFormControlType() in form_autofill_util.cc
- StringToFormControlTypeDiscouraged() in form_field_data.cc
All other changes are iOS boilerplate code or tests.

The change is behind a Finch flag.

Bug: 40283901, 402071086
Change-Id: I59ee08f94a6dafffe062703e97db34132a885a3a
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196749
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1480313}
NOKEYCHECK=True
GitOrigin-RevId: d3c335d89d1d85f0145beb3fc8d22a61a9b8d107