Fix potential crash in ServiceWorkerRegistry::FindRegistrationForClientUrl
This CL targets to fix the following test failures.

https://ci.chromium.org/ui/test/chromium/ninja%3A%2F%2Fandroid_webview%2Ftest%3Awebview_instrumentation_test_apk%2Forg.chromium.android_webview.test.AwQuotaManagerBridgeTest%23testDeleteBrowsingDataClearsServiceWorker?q=V%3Abuilder%3Dandroid-arm64-proguard-rel+V%3Adevice_os%3DQQ1A.191205.008+V%3Adevice_type%3Dwalleye+V%3Aos%3DAndroid+V%3Atest_suite%3Dwebview_instrumentation_test_apk_single_process_mode

In `ServiceWorkerRegistry::FindRegistrationForClientUrl`, the `scopes`
variable is an `std::optional`. When a preflight result was found, the
code attempted to dereference `scopes` to construct a vector without
checking if it contained a value. This could lead to a crash if `scopes`
was `std::nullopt`.

This change adds a check for `scopes.has_value()` before accessing its
contents, preventing the potential crash. It constructs an optional
vector to pass to `DidFindRegistrationForClientUrl`, which correctly
handles the `std::nullopt` case.

Bug: 428130197
Change-Id: I90eb07456ab2d6336901c2ee80b6c303ae07e100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6698612
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481449}
NOKEYCHECK=True
GitOrigin-RevId: 4d727f18a3e3a27a511321f847b915860f5f2246