[Binder] Use #getActiveAdmins for Managed Profile check
Currently, #getDeviceEnterpriseInfoInBackground is making hundreds of
Binder calls on startup (~700 from local traces!) placing heavy stress
on the system. While this is in the background, freeing up the Binder
thread allows it to handle other important tasks instead.

The impact of this can be lessened by iterating through active admins
instead of all packages. This CL introduces an experiment to do that and
brings back the timing histogram to observe the effects of the change.

Bug: 425728303
Change-Id: I3a8c4490148602a3a92a0d01ecfea40266f7ac1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6664758
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Nafis Abedin <nafisabedin@google.com>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1483149}
NOKEYCHECK=True
GitOrigin-RevId: 5ece4ba03dc725f942bc0a51a59926602e318bcb