Add field trial config for OffloadAcceptCHFrameCheck
This CL adds a field trial testing configuration for the
OffloadAcceptCHFrameCheck feature. This will allow starting an
experiment for the feature.

The experiment is enabled on Windows, Mac, Linux, ChromeOS, and Android.

This CL includes the fix on CHECK failure in NavigationURLLoaderImpl,
and test failure in client_hints_browsertest, which was caused by adding
the field trial testing config.

To understand how often it happens, the CL adds a metric for that.

Bug: 406407746, 430254404
Change-Id: I6f07a77db3f41cbe9c5fd24f1d42e893309f2b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6703688
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Victor Tan <victortan@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1485289}
NOKEYCHECK=True
GitOrigin-RevId: 836f8916859e47b2531d4b0f6960937892e0a474