Fully launch tab search IPH when tab search is in toolbar button
Due to how the IPH system uses feature flags, it would be tricky to
mirror the solution used by the toolbar button migration to gate on OS
and US users. Instead this relies on the fact that we check that the
toolbar button migration is true before showing this IPH:
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/views/frame/browser_view.cc;l=3091-3099;drc=285461c1e4957f186a9273ae4db8796ea3c35c88

That way we can just enable this feature and any user who has the
toolbar button would be eligible for the IPH. Technically this will mean
launching it for some users (1% on non-US and ChromeOS) who didn't have
the IPH previously but there shouldn't be any differences there.

Bug: 418779242
Change-Id: I656ca00bef07c6ec677ff3f51d38c81a672c2ab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6785727
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Auto-Submit: Alison Gale <agale@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1492354}
NOKEYCHECK=True
GitOrigin-RevId: 9df594b1918134b8c41a306dfd0da7edf48870ee