Use priority-based TaskRunner in HttpCache::Transaction
This CL updates HttpCache::Transaction to use the new priority-based
task runners, as part of the ongoing effort to migrate the network stack
to the NetworkServiceTaskScheduler.

HttpCache::Transaction was previously identified as a high-risk
component and was excluded from the initial Finch Trial. This change now
enables the scheduler for HttpCache::Transaction for the next finch,
Phase 2 or later, guarded by the `http_cache_transaction` feature
parameter.

Bug: 421051258
Change-Id: Ic311eab5cde986e0deee7f7497d74cb0a7337497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6791128
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1493218}
NOKEYCHECK=True
GitOrigin-RevId: d720139371e6c0e555e30d54b48025a22a150e99