ReloadHiddenTabsWithCrashedSubframes only reloads for active subframes
We should not reload the WebContents when a subframe in back-forward
cache is crashed.

Prerendering subframes do not enter the path as the comment says.

For back-forward cache subframes, the cached page which contains the
crashed frame is evicted at RenderFrameHostImpl::RenderProcessGone().

Bug: 434100501
Change-Id: I686f2df63d7b6614ff14a477034112ac9d88029a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6785662
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>
Cr-Commit-Position: refs/heads/main@{#1493225}
NOKEYCHECK=True
GitOrigin-RevId: b6c3190bcef841f0fb32cb8e8fb3d688b66191d6