Add field trial config for DiskCacheBackendExperiment
This change adds the field trial testing configuration for the
`DiskCacheBackendExperiment`. This allows testing the newly added `sql`
backend options on the CI bots.

This is necessary to get test coverage for the new backend options and
to start a field trial tests.

Bug: 422065015
Change-Id: I3efb717ea5b4b5989a48c80b7c390b39f68ee7a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6787188
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1493359}
NOKEYCHECK=True
GitOrigin-RevId: df398ed73c9b97e1b0e6cd2fd697f550c907706a