Send CSPP ping on Clipboard copy API call
This CL makes CSPP listen to the `OnTextCopiedToClipboard` web contents
observer. The handler will run whenever there is text copied to the
clipboard. It is API agnostic, so it will run regardless of if text was
copied via `navigator.clipboard` or `document.execCommand('copy')`.
However, it will only run if text is copied to the clipboard (as opposed
to images or binary data).

The new behavior is gated by a feature that stops processing after
preclassification. It also has a parameter that determines the sampling
rate if a site matches the HC allowlist.

Change-Id: If930378181955ca52f833a0494b302f47be995f3
Bug: 434217994
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6775701
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Seung Jae Lim <andysjlim@chromium.org>
Commit-Queue: Andrew Liu <liu@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1495005}
NOKEYCHECK=True
GitOrigin-RevId: c93137ed8335c0e4437c1361ec90efee81f18a55