Use priority-based TaskRunner in HttpCache
This CL updates HttpCache to use the new priority-based task runners, as
part of the ongoing effort to migrate the network stack to the
NetworkServiceTaskScheduler.

While an `HttpCache` instance itself doesn't map directly to a single
`URLRequest`, each `HttpCache::Transaction` has a priority. This change
uses the highest request priority among all transactions associated with
an active entry to set the task priority when posting a task to process
the entry.

`HttpCache` was previously identified as a high-risk component and
excluded from the initial Finch trial. This change enables the scheduler
for `HttpCache` for the next Finch trial (Phase 2 or later), controlled
by the `http_cache` feature parameter.

Bug: 421051258
Change-Id: I539f7cfe7895651d7e184b91f6e37d2c1c3fe153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6796711
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496657}
NOKEYCHECK=True
GitOrigin-RevId: 30ecce9ca0f5a103e35f2b6aae00c1fc1f887211