Remove the vpython spec file for generate_buildbot_json_unittest.py
This will cause the test to use the root .vpython3 file instead, which
ensures it tracks the version of python and packages used by the
majority of python within chromium/src. As it is the unittest is
currently run with python 3.8 instead of 3.11.

Change-Id: I7c7c5fc20d906d88e690f16c8e8d4988a1d9b2ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6828101
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498474}
NOKEYCHECK=True
GitOrigin-RevId: 34f8816c3a90ac3fbad7842fc360dd9f3935d7f4