Remove page action icon for Click to Call
This CL aims to remove Click To Call page action icon from being
rendered as part of the attempt to migrate in-use page actions to the
new framework and delete the rest.

Click to Call feature has been disabled for a long time and is planned
to be removed soon. The Device Picker bubble is already anchored
to RHS of the Omnibox.

Fixed: 376283367
Change-Id: Ic259759c3ec4375d1cd95ac490ef6700d26cfe65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6838976
Commit-Queue: Atharv Maan <atharvmaan@chromium.org>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Reviewed-by: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501385}
NOKEYCHECK=True
GitOrigin-RevId: c70e15e12aa0d2a108b5426fef63435f0f5d2058