[content] Add flag to keep isolates at best effort priority in hidden renderers.
The SetIsolatesPriority feature uses the more granular process priority
instead of renderer visibility to set isolate priority (a process
hosting small visible frames that didn't receive input has
user-visible priority, whereas one with a large visible frame has
user-blocking priority, but both are visible).

However, because a hidden renderer can have user-blocking priority (e.g.
if it plays audio), this has the effect of elevating the priority of
some isolates. This is orthogonal to our goal of reducing the isolates
priority for renderers hosting visible but unimportant frames. This CL
adds a IsolatesPriorityBestEffortWhenHidden flag which forces isolates
priority in hidden renderers to best-effort. We'll use it it a new
experiment with unimportant frames policy. We keep it as a separate
feature from IsolatesPriorityUseProcessPriority, to allow evaluating the
impact of disabling it if experiment results suggests that it's worth
it.

Bug: 40942028, 350698335
Change-Id: Ie58c8b67cd0a15713bffd2ae0ef090a01f6b4f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6828565
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1501888}
NOKEYCHECK=True
GitOrigin-RevId: 9f180d1b048570a16982fa60217421e39947786d