Add a flag to load native library early (part 2)
A previous CL added a flag called LoadNativeEarly. This CL adds a
boolean parameter called LoadNativeEarlyConcurrentLoad.
When this param is true, we first load the chrome split
asynchronously in a background thread, then load the native library
synchronously in the main thread. When this param is false,
we first load the native library synchronously in the main thread,
and then when it finishes, we start loading the chrome split
(this is the behavior introduced in the previous CL).

Bug: 437109332
Change-Id: Ie6bb1a52e92f0a2dff7d1e19a47ab57c58678c7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6850266
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>
Cr-Commit-Position: refs/heads/main@{#1502713}
NOKEYCHECK=True
GitOrigin-RevId: 42109860e226fa9539fe0341c291880c0fb7abb0